Removed more job poll spam, migrated some older tests
