+++ src/biokbase/narrative/tests/logproxy.conf
