+++ src/biokbase/narrative/tests/test_func_info.py
