+++ src/biokbase/narrative/tests/test_log_proxy.py
